+++ LAB3/Core/Src/fsm_automatic.c
@@ -12,7 +12,7 @@ void fsm_automatic_run(){
 		 clear();
 
 		 status = AUTO_RED;
-		 set_Timer1(red_duration*100);
+		 set_Timer1((int)(red_duration*100));
 
 		 break;
 	 case AUTO_RED:
@@ -20,7 +20,7 @@ void fsm_automatic_run(){
 		 if(timer1_flag ==1)
 		 {
 			 status = AUTO_GREEN;
-			 set_Timer1(green_duration*100);
+			 set_Timer1((int)(green_duration*100));
 		 }
 		 if(isButton1Pressed()==1)
 		 {
@@ -35,7 +35,7 @@ void fsm_automatic_run(){
 		 {
 
 			 status = AUTO_YELLOW;
-			 set_Timer1(yellow_duration*100);
+			 set_Timer1((int)(yellow_duration*100));
 		 }
 		 if(isButton1Pressed()==1)
 		 {
@@ -50,7 +50,7 @@ void fsm_automatic_run(){
 		 if(timer1_flag ==1)
 		 {
 			 status = AUTO_RED;
-			 set_Timer1(red_duration*100);
+			 set_Timer1((int)(red_duration*100));
 		 }
 		 if(isButton1Pressed()==1)
 		 {
